+++ gradle-plugin/src/main/kotlin/io/github/khiaroslav/stringveil/gradle/StringVeilGradlePlugin.kt
@@ -39,6 +39,10 @@ public class StringVeilGradlePlugin : KotlinCompilerPluginSupportPlugin {
                     value = (kotlinCompilation.platformType == KotlinPlatformType.androidJvm)
                         .toString(),
                 ),
+                SubpluginOption(
+                    key = "failOnSecretLikeLiterals",
+                    value = extension.failOnSecretLikeLiterals.getOrElse(false).toString(),
+                ),
             )
         }